% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{add_silhouette_to_dta}
\alias{add_silhouette_to_dta}
\title{Add silhouette information on the the dataset}
\usage{
add_silhouette_to_dta(dta, tbl_silhouette)
}
\arguments{
\item{dta}{The original data}

\item{tbl_silhouette}{A dataframe. Output of the \code{get_sil_widths}}
}
\value{

}
\description{
It adds not only the `sil_width` but also the cluster membership
and the neighbor cluster
}
