% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_cluster.R
\name{analysis_cluster}
\alias{analysis_cluster}
\title{Clustering patients based on the disease}
\usage{
analysis_cluster(input, clusterSubject, Model, clustermax, itermax, mc)
}
\arguments{
\item{input}{the dataframe having patientd ID and phenotype information}

\item{clusterSubject}{A vector of phenotypes or ICD-9 codes of disease of interest that is used to cluster the patients}

\item{clustermax}{Maximum threshold of number of clusters that is defined for mode-based cluster function}

\item{itermax}{Maximum number of iteration to analysis}

\item{mc}{number of cores for cluster Analysis}
}
\description{
Clustering patients based on the disease
}
\examples{
example1 <- analysis_cluster (
                    input                =ASDGI patients,
                    clusterSubject       =Gastrointestinal phenotypes,
                    clustermax            =5,
                    itermax              =200,
                    mc                   =2)

}
