% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{as_tbl_sep_matrix}
\alias{as_tbl_sep_matrix}
\title{Tbl separation matrix}
\usage{
as_tbl_sep_matrix(sep_matrix)
}
\arguments{
\item{cluster_stats}{A list that contains
- separation.matrix
- cluster.number
Usually a results of the fpc::cluster.stats()}
}
\description{
Tbl separation matrix
}
