% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_instability.R
\name{bootstrapInstability}
\alias{bootstrapInstability}
\title{Fang and Wang's instability measure}
\usage{
bootstrapInstability(X, maxK, B = 50, clusterAlg = myKmean,
  similarity = adj.rand.index, verbose = TRUE, ...)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{maxK}{maximum number of clusters to evaluate}

\item{B}{number of resampling iterations}

\item{clusterAlg}{clustering algorithm. Its output must be a list containing attributs "cluster" and "predict".
For more details, check the formatting of function \code{\link{myKmean}}.}

\item{similarity}{function measuring the similarity between two partitions.}

\item{verbose}{logical. If TRUE, plots the evolution of the algorithm}

\item{...}{additional parameters for the clustering algorithm}
}
\value{
List with 3 components:
\describe{
\item{\preformatted{inst_mean}}{vector containing the mean instability measure for 2 to maxK cluster number}
\item{\preformatted{kopt}}{the optimal number of clusters}
\item{\preformatted{instability}}{matrix containing the instability measures for all cluster number and all subsampling iterations.}
}
}
\description{
This method is based on the algorithm developped by Fang and Wang but with more choice regarding the instability measure.
Their measure is equivalent to 1 - rand.index, here one can chose any normalized similarity measure and the instability will be 1 - similarity.
For each number of clusters, several pair of bootstrap subsambles are selected and the instability measure is computed from the clustering of these pairs.
The optimal number of clusters is the value for which the instability is the lowest.
}
\references{
Fang, Y. and Wang, J. (2012). Selection of the number of clusters via the bootstrap method. Computational Statistics  Data Analysis, 56:468-477.
}
