% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data.R
\name{build_data}
\alias{build_data}
\title{Data preparation function}
\usage{
build_data(observationData, conceptData, demogData, minimumAge = 0,
  maximumAge = 18)
}
\arguments{
\item{observationData}{Data frame containing Observation data file}

\item{conceptData}{Data frame containing Concept data file}

\item{demogData}{Data frame containing Demogeraphic data file}

\item{minimumAge}{The minimum age for patients to be included in the analysis.
By default it is set to 0.}

\item{maximumAge}{The maximum age for patients to be included in the analysis.
By default it is set to 18.}
}
\description{
Data preparation function
}
