% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{calc_diss_matrix}
\alias{calc_diss_matrix}
\title{Calculate the dissimilarity matrix}
\usage{
calc_diss_matrix(dta, metric = c("euclidean", "manhattan", "gower"))
}
\arguments{
\item{dta}{A dataframe. The data}
}
\description{
Calculate the dissimilarity matrix
}
\details{
Uses the \code{\link{cluster::daisy}} function. This function
needs the character columns as factors to properly function
}
