% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clest.R
\name{clest}
\alias{clest}
\title{Clest: A prediction-based resampling method for estimating the number of clusters in a dataset}
\usage{
clest(X, maxK, clusterAlg = myKmean, similarity = adj.rand.index,
  pmax = 0.05, dmin = 0.05, B = 50, B0 = 20, rho = 0.6,
  verbose = TRUE, ...)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{maxK}{maximum number of clusters to evaluate.}

\item{clusterAlg}{clustering algorithm. Its output must be a list having a compoment "cluster" containing the assignation of each observation.
For more details, check the formatting of function \code{\link{myKmean}}.}

\item{similarity}{function measuring the similarity between two partitions.}

\item{pmax}{threshold for the p-value}

\item{dmin}{threshold for the d-value}

\item{B}{number of resampling iterations}

\item{B0}{number of reference datasets to generate}

\item{rho}{proportion of the train/test set}

\item{verbose}{logical, if TRUE, plots the evolution of the algorithm}

\item{...}{additional parameters for the clustering algorithm}
}
\value{
list of 3 attributes:
\describe{
\item{\preformatted{p}}{vector of p-values}
\item{\preformatted{d}}{vector of d-values}
\item{\preformatted{kopt}}{optimal number of clusters}
}
}
\description{
This method measures the similarity of two clustering computed on two non-overlapping subsamples of the dataset.
The obtained measures are then compared with other similarity measures taken on generated datasets.
Those generated datasets are computed as in the gap statistics method (Tibshirani et al, 2001).
}
\references{
\itemize{
\item{Dudoit, S. and Fridlyand, J. (2002). A prediction-based resampling method for estimating the number of clusters in a dataset. Genome Biology, 3(7):research0036.1.
\url{https://doi.org/10.1186/gb-2002-3-7-research0036}}
\item{Tibshirani, R., Walther, G., and Hastie, T. (2001). Estimating the number of clusters in a data set via the gap statistic.Journal of the Royal Statistical Society Series B, 63:411-423.}}
}
