% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_plot.R
\name{comorbidity_plot}
\alias{comorbidity_plot}
\title{Comorbidity plot of clusters}
\usage{
comorbidity_plot(L, pval)
}
\arguments{
\item{L}{The list of different clusters}

\item{pval}{the selected pvalue for defining the significant of each comorbidity}
}
\description{
Comorbidity plot of clusters
}
\examples{
example1 <- comorbidity_plot(
                      L    =list(subgroup1,subgroup2,subgroup3)
                      pval= 0.05
                    )

 @export comorbidity_plot
}
