% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{data_subset}
\alias{data_subset}
\title{Data Subgroups definition}
\usage{
data_subset(input, ICD_Disease, ICD_Comorbidity, varnoInterest)
}
\arguments{
\item{input}{Main Data frame containing the disease and comorbidity}

\item{ICD_Disease}{vector of ICD-9 codes (character) for selecting the disease}

\item{ICD_Comorbidity}{ICD-9 codes for selecting the comorbidity}

\item{varnoInterest}{ICD-9 categorical codes}
}
\value{
A list of Data frames containing the Disease data, Comorbidity data and Disease+comorbidity data
}
\description{
Data Subgroups definition
}
\examples{

example1 <- data_subset(
              input              = MasterData,
              ICD_Disease        =c(299.0,299.8,299.9),
              ICD_Comorbidity    =c(530,580,787,788)
              varnoInterest      = c("V","E"),

              )
}
