% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_selection.R
\name{disease_selection}
\alias{disease_selection}
\title{A function to select a subset of patients  diagnosed with a specific disease}
\usage{
disease_selection(input, varInterest, varnoInterest)
}
\arguments{
\item{input}{Clincial records with Diagnosis}

\item{varInterest}{The subset of codes that are used to define the disease of interest}

\item{varnoInterest}{The subset of codes that are not used to define the disease of interest}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
An object of class \code{data.frame} with the query output.
}
\description{
A function to select a subset of patients  diagnosed with a specific disease
}
