% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{facet_panel_dimensions}
\alias{facet_panel_dimensions}
\title{Extract the panel dimensions in a faceted ggplot}
\usage{
facet_panel_dimensions(p)
}
\arguments{
\item{p}{A ggplot2 object}
}
\value{
A named list of the number of $rows and $cols
}
\description{
This function extracts the number if rows and columns of the panel
in a faceted ggplot. It can be used to calculate the \code{width} and \code{height}
of a \code{shiny::renderPlot()}
}
\details{
Some details here :)
}
\examples{
library(ggplot2)
p1 <- ggplot(mtcars, aes(mpg, wt)) + facet_wrap(~ gear )
facet_panel_dimensions(p1)
p2 <- ggplot(mtcars, aes(mpg, wt)) + facet_grid(cyl ~ gear )
facet_panel_dimensions(p2)
}
