% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{get_sil_widths}
\alias{get_sil_widths}
\title{Get the silhouette widths}
\usage{
get_sil_widths(obj, diss_matrix)
}
\arguments{
\item{obj}{A clustering object}

\item{diss_matrix}{A dissimilarity matrix, of class `dist`}
}
\value{
A tibble with the observation id, the cluster membership, 
the cluster neighbor and the silhouette width
}
\description{
This function extracts or calculates the silhouette widths from the clustering objects
}
\details{
\code{cluster::pam} retains a table of sil widths in obj$silinfo$widths
For the other cluster methods we calculate them ourselves using
cluster::silhouette    

The clustering `obj` must have a named `cluster` vector. Named with the observation id
and the value the observation cluster membership
}
