% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{get_var_labels}
\alias{get_var_labels}
\title{Adapted from 
https://github.com/larmarange/labelled/blob/main/R/var_label.R}
\usage{
get_var_labels(x, unlist = FALSE)
}
\description{
Adapted from 
https://github.com/larmarange/labelled/blob/main/R/var_label.R
}
