% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{gg_density_plot}
\alias{gg_density_plot}
\title{Create a density plot}
\usage{
gg_density_plot(dta)
}
\arguments{
\item{dta}{The data}
}
\value{
A ggplot. The result is a faceted plot (facet: each variable)
}
\description{
This function creates a faceted density plot to demonstrate the 
distribution of the variables within each cluster.
}
