% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{gg_separation_matrix}
\alias{gg_separation_matrix}
\title{Plot a separation matrix}
\usage{
gg_separation_matrix(sep_matrix)
}
\arguments{
\item{sep_matrix}{An nxn matrix.}
}
\value{

}
\description{
This function plots a tile plot of the separation matrix of a cluster solution
}
\details{
Assume n clusters, the separation matrix is a `n x n` table, with 
cluster separation values for each pair of clusters. This is the result of 
fpc::cluster.stats()$separation.matrix
}
