% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gplot.R
\name{Gplot}
\alias{Gplot}
\title{Gplot}
\usage{
Gplot(input_data, clusterSubject, n)
}
\arguments{
\item{input_data}{The dataframe used to the prevalence of data}

\item{clusterSubject}{vector of disease symptoms that are used fro clustering}

\item{n}{the minimum prevalence threshold}
}
\description{
Gplot
}
\examples{
example1 <- Gplot(
                     input_data    =GIAut,
                      clusterSubject  =clusterSubject,
                      n               =5
                    )

 @export Gplot
}
