% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter-center.R
\name{gupta}
\alias{gupta}
\title{Last Leap and Last Major Leap}
\usage{
gupta(X, maxK, clusterAlg = myKmean, verbose = FALSE, ...)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{maxK}{maximum number of cluster to evaluate}

\item{clusterAlg}{clustering algorithm. Its output must be a list having an attribute "centers" containing the centers of each cluster.
For more details, check the formatting of function \code{\link{myKmean}}.}

\item{verbose}{logical, if TRUE it plot the evolution of the algorithm}

\item{...}{additional parameters for the clustering algorithm}
}
\value{
list with 5 compoments:
\describe{
\item{\preformatted{d}}{vector of the minimum inter-center distance for each number of cluster}
\item{\preformatted{ll}}{vector containing the relative difference between d(k) and d(k+1)}
\item{\preformatted{ll_kopt}}{optimal number of clusters according to the last leap method}
\item{\preformatted{lml}}{vector indicating wether there is a major gap between values d(k) and d(k+1)}
\item{\preformatted{lml_kopt}}{optimal number of clusters according to the last major leap method}
}
}
\description{
This method first computes the minimum distance between two cluster centers d.
Then, it gets the optimal number of clusters according to the last leap method and the last major leap method.
}
\references{
Gupta, A., Datta, S., and Das, S. (2018). Fast automatic estimation of the number of clusters from the minimum inter-center distancefor k-means clustering.Pattern Recognition Letters, 116.
}
