% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolesnikov.R
\name{kolesnikov}
\alias{kolesnikov}
\title{Quantatization error modeling}
\usage{
kolesnikov(X, maxK, clusterAlg = myKmean)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{maxK}{maximum number of clusters to evaluate}

\item{clusterAlg}{clustering algorithm. Its output must be a list containing parameters "cluster" and "center".
For more details, check the formatting of function \code{\link{myKmean}}.}
}
\value{
list having 3 attributes:
\describe{
\item{\preformatted{kopt}}{optimal number of clusters}
\item{\preformatted{PCF}}{vector of score}
\item{\preformatted{a}}{value of the exponent \emph{a}}
}
}
\description{
This method is based on the within cluster variance W. The idea is to estimate a parameter \emph{a} and to find the minimum of this function : \eqn{W(k)k^a}{W(k)k^a}.
}
\references{
Kolesnikov, A., Trichina, E., and Kauranne, T. (2015). Estimating the number of clusters in a numerical data set via quantizationerror modeling.Pattern Recognition, 48:941-952.
}
