% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelExplorer.R
\name{ModelExplorer}
\alias{ModelExplorer}
\title{Stability measure for finding the number of clusters}
\usage{
ModelExplorer(X, maxK, similarity = adj.rand.index,
  clusterAlg = myKmean, rho = 0.8, B = 100, verbose = FALSE, ...)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{maxK}{maximum number of clusters to evaluate.}

\item{similarity}{function measuring the similarity between two partitions.}

\item{clusterAlg}{clustering algorithm. Its output must be a list having a compoment "cluster" containing the assignation of each observation.
For more details, check the formatting of function \code{\link{myKmean}}.}

\item{rho}{numeric between 0 and 1. Proportional size of the subsamples.}

\item{B}{Number of resampling iterations.}

\item{verbose}{logical. If TRUE, it plot the evolution of the algorithm.}

\item{...}{additional parameters for the clustering algorithm.}
}
\value{
Matrix of size maxK-1 x B containing the stability measure for each iteration.
}
\description{
Model Explorer measures the stability of a given clustering method on subsambles of the original dataset.
It plots the cumulative distribution of the stability measure. The more the distribution is concentrated on the right, the better.
}
\references{
Ben-Hur, A., Elisseeff, A., and Guyon, I. (2002). A stability based method for discovering structure in clustered data.Pacific Symposium on Biocomputing. Pacific Symposium on Biocomputing, 2002:6-17
}
