% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{myHierarchicalClustering}
\alias{myHierarchicalClustering}
\title{Hierarchical clustering}
\usage{
myHierarchicalClustering(X, k, distance = "euclidean",
  linkage = "average")
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{k}{number of clusters}

\item{distance}{character for the type of distance. Default is "euclidean"}

\item{linkage}{"single", "average" or "complete"}
}
\value{
list of 3 components:
\describe{
\item{\preformatted{cluster}}{vector of integer between 1 and k containing the allocation of each point}
\item{\preformatted{centers}}{matrix (d x k) of the centers of each cluster}
\item{\preformatted{predict}}{function predicting to which cluster an observation belongs. The prediciton is done with the k-nearest neighbours function.}
}
}
\description{
This function performs hierarchical clustering in either average, complete or single linkage.
}
