% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{myKmean}
\alias{myKmean}
\title{K-means clustering}
\usage{
myKmean(X, k)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{k}{number of clusters}
}
\value{
list of 3 components:
\describe{
\item{\preformatted{cluster}}{vector of integer between 1 and k containing the allocation of each point}
\item{\preformatted{centers}}{matrix (d x k) of the centers of each cluster}
\item{\preformatted{predict}}{function predicting to which cluster an observation belongs. The input can be a single observation vector or a matrix of several observations. The assigned cluster is the one with the closest center to the given observation.}
}
}
\description{
This function is a wrapper to \code{\link[stats:kmeans]{kmeans}} in the package stats.
}
