% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_algorithms.R
\name{myPam}
\alias{myPam}
\title{Partitioning around medoids}
\usage{
myPam(X, k, ...)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features}

\item{k}{number of clusters}

\item{...}{additional parameter for pam function}
}
\value{
list of 3 components:
\describe{
\item{\preformatted{cluster}}{vector of integer between 1 and k containing the allocation of each point}
\item{\preformatted{centers}}{matrix (d x k) of the centers of each cluster}
\item{\preformatted{predict}}{function predicting to which cluster an observation belongs. The prediciton is done with the k-nearest neighbours function.}
}
}
\description{
This function is a wrapper to \code{\link[cluster:pam]{pam}} in the package cluster.
}
