% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{3D dataset visualization}
\usage{
plot3D(X, label = NULL)
}
\arguments{
\item{X}{data matrix or data frame of size n x d, n observations and d features. d must be superior or equal to 3.}

\item{label}{array of size n containing each cluster assignation for each observation.}
}
\description{
Function for plotting the 3 principal components of a dataset. One can also add color labels to visualise cluster assignations.
}
