% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_file_fct_utils.R
\name{process_file}
\alias{process_file}
\title{Process the uploaded file}
\usage{
process_file(dta)
}
\arguments{
\item{dta}{A data.frame The data read by the `read_file()`}
}
\value{
A tibble of the dataset
}
\description{
This function performs some processing on the uplaoded file
before it is used by the app
}
