% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_significant.R
\name{statistical_significant}
\alias{statistical_significant}
\title{Statistical Chi-square test to define the significant comorbidities associated to different subgroups of data}
\usage{
statistical_significant(DiseaseData1, DiseaseData2, pval, ORmin, countmin)
}
\arguments{
\item{DiseaseData1}{The primary data frame for defining its associated comorbidities (PAITNET_ID and Pentoype is used)}

\item{DiseaseData2}{The secondary data frame for comparing the association with primary data frame}

\item{pval}{The original pvalue before bonferroni multiple testing correction}

\item{ORmin}{The minimul acceptable value for Odd Ratio}

\item{countmin}{The minimum acceptable count of patients from both groups having the same comorbidity}
}
\description{
Statistical Chi-square test to define the significant comorbidities associated to different subgroups of data
}
