% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{vars_of_type}
\alias{vars_of_type}
\title{Find variables within a dataset
 
 This function finds variables of a specific type within a dataset
 
 Adapted (i.e. stolen) from 
 https://mastering-shiny.org/scaling-modules.html#case-study-selecting-a-numeric-variable
 The \code{type} argument will lead to a function - Usually \code{is.numeric} or \code{is.character}
 @param data A dataframe
 @param type The type of the variable. One of c("numeric", "character", "factor")
 @export
 @return A character vector of variable names}
\usage{
vars_of_type(data, type = c("numeric", "character", "factor"))
}
\description{
Find variables within a dataset
 
 This function finds variables of a specific type within a dataset
 
 Adapted (i.e. stolen) from 
 https://mastering-shiny.org/scaling-modules.html#case-study-selecting-a-numeric-variable
 The \code{type} argument will lead to a function - Usually \code{is.numeric} or \code{is.character}
 @param data A dataframe
 @param type The type of the variable. One of c("numeric", "character", "factor")
 @export
 @return A character vector of variable names
}
