% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{wrap_title}
\alias{wrap_title}
\title{Wrap titles}
\usage{
wrap_title(x, ...)
}
\description{
Function to wrap titles, so they show completely when saving plot in ggplot
}
\details{
Stolen from https://github.com/Public-Health-Scotland/scotpho-profiles-tool/blob/master/shiny_app/global.R
}
