% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{as_tbl_sep_matrix}
\alias{as_tbl_sep_matrix}
\title{Turn separation matrix to a tibble}
\usage{
as_tbl_sep_matrix(sep_matrix)
}
\arguments{
\item{sep_matrix}{A separation matrix, usually from fpc::cluster.stats()}
}
\description{
Turn separation matrix to a tibble
}
