% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data.R
\name{build_data}
\alias{build_data}
\title{A function in order to create the data set for the analysis}
\usage{
build_data(input, codeRef, RefTable, refPath, minimumAge = 0,
  maximumAge = 18)
}
\arguments{
\item{input}{Data frame containing clinical data file for analysis}

\item{codeRef}{XXXX}

\item{RefTable}{YYYY}

\item{refPath}{Contains the path and name of the phenotype mapping file}

\item{minimumAge}{The minimum age for patients to be included in the analysis.
By default it is set to 0.}

\item{maximumAge}{The maximum age for patients to be included in the analysis.
By default it is set to 18.}
}
\description{
A function in order to create the data set for the analysis
}
