% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Plot of clusters}
\usage{
cluster_plot(L, ClusterSubject, n)
}
\arguments{
\item{L}{The list of different clusters}

\item{ClusterSubject}{vector of disease symptoms that are used fro clustering}

\item{n}{the minimum prevalence threshold}
}
\description{
Plot of clusters
}
\examples{
example1 <- cluster_plot(
                      L               =list(subgroup1,subgroup2,subgroup3)
                      ClusterSubject  =GI$phenotypes
                      n               =5
                    )

 @export cluster_plot
}
