% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{data_subset}
\alias{data_subset}
\title{Data Subgroups definition}
\usage{
data_subset(input, ICD_Disease, ICD_Comorbidity)
}
\arguments{
\item{input}{Main Data frame containing the disease and comorbidity}

\item{ICD_Disease}{vector of ICD-9 codes (character) for selecting the disease}

\item{ICD_Comorbidity}{ICD-9 codes used for defining the comorbidity interval}
}
\value{
A list of Data frames containing the Disease data, comorbidity data and Disease+comorbidity data
}
\description{
This function is used to define the subsets of disease and comorbidity and disease+comorbidity
}
\examples{

example1 <- data_subset(
              input              = MasterData,
              ICD_Disease        =c(299.0,299.8,299.9),
              ICD_Comorbidity    =c(530,580,787,788)


              )
output is list of 3 elements
output[[1]]= dataset of Disease
output[[2]]=dataset of comorbidity
output[[3]]=dataset of Disease+comorbidity


}
