% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_file_fct_utils.R
\name{read_file}
\alias{read_file}
\title{Read the uploaded file}
\usage{
read_file(path)
}
\arguments{
\item{path}{String length 1. The path of the dataset}
}
\value{
A tibble of the dataset
}
\description{
This function reads the uploaded file.
}
\details{
The \code{read_file} function currently reads \code{Excel}, 
\code{csv}, \code{tsv}, and \code{SPSS} files. If none of the above file type, 
then an \code{error} message is propagated throughout
}
