% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_significant.R
\name{statistical_significant}
\alias{statistical_significant}
\title{Statistical Significant test}
\usage{
statistical_significant(DiseaseData1, DiseaseData2, pval, ORmin, countmin, pt)
}
\arguments{
\item{DiseaseData1}{The primary data frame for defining its associated comorbidities (PAITNET_ID and Phenotype is used)}

\item{DiseaseData2}{The secondary data frame for comparing the association with primary data frame}

\item{pval}{The original p-value before bonferroni multiple testing correction}

\item{ORmin}{The minimum acceptable value for Odd Ratio}

\item{countmin}{The minimum acceptable count of patients from both groups having the same comorbidity}

\item{pt}{The confidence interval probability}
}
\description{
This function is used to test significance of comorbidities between two subsets of data.
}
